+++ Elements/src/Serialization/glTF/GltfExtensions.cs
@@ -86,7 +86,7 @@ namespace Elements.Serialization.glTF
                     if (SaveGlb(model, path, out errors, drawEdges))
                     {
                         return;
-                    }
+                }
                     // Else fall through to produce an empty GLTF.
                 }
                 else
@@ -225,6 +225,22 @@ namespace Elements.Serialization.glTF
                             {"glossinessFactor", material.GlossinessFactor}
                         }}
                     };
+                }
+
+                if (material.EdgeDisplaySettings != null)
+                {
+                    if (gltfMaterial.Extensions == null)
+                    {
+                        gltfMaterial.Extensions = new Dictionary<string, object>();
+                    }
+                    if (!gltf.ExtensionsUsed.Contains("HYPAR_materials_edge_settings"))
+                    {
+                        gltf.ExtensionsUsed = new List<string>(gltf.ExtensionsUsed) { "HYPAR_materials_edge_settings" }.ToArray();
+                    }
+                    gltfMaterial.Extensions.Add("HYPAR_materials_edge_settings", new Dictionary<string, object>{
+                        {"lineWidth", material.EdgeDisplaySettings.LineWidth},
+                        {"widthMode", (int)material.EdgeDisplaySettings.WidthMode},
+                    });
                 }
 
                 var textureHasTransparency = false;